+++ src/backend/data/ChangeSet.ts
@@ -1,6 +1,7 @@
 import {Config} from 'alinea/core/Config'
 import {META_KEY, createRecord} from 'alinea/core/EntryRecord'
 import {EntryPhase} from 'alinea/core/EntryRow'
+import {Graph} from 'alinea/core/Graph'
 import {
   ArchiveMutation,
   CreateMutation,
@@ -16,8 +17,10 @@ import {
   RemoveEntryMutation,
   UploadMutation
 } from 'alinea/core/Mutation'
-import {EntryUrlMeta, Type} from 'alinea/core/Type'
+import {Query} from 'alinea/core/Query'
+import {Type} from 'alinea/core/Type'
 import {Workspace} from 'alinea/core/Workspace'
+import {MediaFile} from 'alinea/core/media/MediaTypes'
 import {join} from 'alinea/core/util/Paths'
 import {JsonLoader} from '../loader/JsonLoader.js'
 
@@ -68,22 +71,7 @@ const decoder = new TextDecoder()
 const loader = JsonLoader
 
 export class ChangeSetCreator {
-  constructor(public config: Config) {}
-
-  entryLocation(
-    {locale, parentPaths, path, phase}: EntryUrlMeta,
-    extension: string
-  ) {
-    const segments = (locale ? [locale] : [])
-      .concat(
-        parentPaths
-          .concat(path)
-          .map(segment => (segment === '' ? 'index' : segment))
-      )
-      .join('/')
-    const phaseSegment = phase === EntryPhase.Published ? '' : `.${phase}`
-    return (segments + phaseSegment + extension).toLowerCase()
-  }
+  constructor(protected config: Config, protected graph: Graph) {}
 
   editChanges({previousFile, file, entry}: EditMutation): Array<Change> {
     const type = this.config.schema[entry.type]
@@ -160,10 +148,34 @@ export class ChangeSetCreator {
     ]
   }
 
-  removeChanges({file}: RemoveEntryMutation): Array<Change> {
+  async removeChanges({
+    entryId,
+    file
+  }: RemoveEntryMutation): Promise<Array<Change>> {
     if (!file.endsWith(`.${EntryPhase.Archived}.json`)) return []
+    const {workspace, files} = await this.graph.preferPublished.get(
+      Query.whereId(entryId).select({
+        workspace: Query.workspace,
+        files: Query.children<typeof MediaFile>(undefined!, 999).where(
+          Query.type.is('MediaFile')
+        )
+      })
+    )
+    const mediaDir =
+      Workspace.data(this.config.workspaces[workspace])?.mediaDir ?? ''
+    const removeFiles: Array<Change> = files.map(file => {
+      const binaryLocation = join(mediaDir, file.location)
+      return {
+        type: ChangeType.Delete,
+        file: binaryLocation
+      }
+    })
     return [
+      // Remove any media files in this location
+      ...removeFiles,
+      // Remove entry
       {type: ChangeType.Delete, file},
+      // Remove children
       {
         type: ChangeType.Delete,
         file: file.slice(0, -`.${EntryPhase.Archived}.json`.length)
@@ -224,7 +236,7 @@ export class ChangeSetCreator {
     return [{type: ChangeType.Delete, file: mutation.file}, removeBinary]
   }
 
-  mutationChanges(mutation: Mutation): Array<Change> {
+  async mutationChanges(mutation: Mutation): Promise<Array<Change>> {
     switch (mutation.type) {
       case MutationType.Edit:
         return this.editChanges(mutation)
@@ -251,9 +263,10 @@ export class ChangeSetCreator {
     }
   }
 
-  create(mutations: Array<Mutation>): ChangeSet {
-    return mutations.map(meta => {
-      return {changes: this.mutationChanges(meta), meta}
-    })
+  async create(mutations: Array<Mutation>): Promise<ChangeSet> {
+    const res = []
+    for (const meta of mutations)
+      res.push({changes: await this.mutationChanges(meta), meta})
+    return res
   }
 }