+++ src/dashboard/view/entry/EntryHeader.tsx
@@ -79,6 +79,7 @@ export function EntryHeader({editor, editable = true}: EntryHeaderProps) {
   const previewRevision = useAtomValue(editor.previewRevision)
   const isActivePhase = editor.activePhase === selectedPhase
   const isMediaFile = editor.activeVersion.type === 'MediaFile'
+  const isMediaLibrary = editor.activeVersion.type === 'MediaLibrary'
   const hasChanges = useAtomValue(editor.hasChanges)
   const currentTransition = useAtomValue(editor.transition)?.transition
   const untranslated = locale && locale !== editor.activeVersion.locale
@@ -100,9 +101,16 @@ export function EntryHeader({editor, editable = true}: EntryHeaderProps) {
   const publishArchived = useSetAtom(editor.publishArchived)
   const deleteArchived = useSetAtom(editor.deleteArchived)
   const deleteFile = useSetAtom(editor.deleteFile)
+  const deleteMediaLibrary = useSetAtom(editor.deleteMediaLibrary)
   const queryClient = useQueryClient()
   function deleteFileAndNavigate() {
     return deleteFile()?.then(() => {
+      queryClient.invalidateQueries('explorer')
+      navigate(nav.root(entryLocation))
+    })
+  }
+  function deleteMediaLibraryAndNavigate() {
+    return deleteMediaLibrary()?.then(() => {
       queryClient.invalidateQueries('explorer')
       navigate(nav.root(entryLocation))
     })
@@ -160,6 +168,15 @@ export function EntryHeader({editor, editable = true}: EntryHeaderProps) {
           <DropdownMenu.Item
             className={styles.root.action()}
             onClick={deleteFileAndNavigate}
+          >
+            Delete
+          </DropdownMenu.Item>
+        </>
+      ) : isMediaLibrary ? (
+        <>
+          <DropdownMenu.Item
+            className={styles.root.action()}
+            onClick={deleteMediaLibraryAndNavigate}
           >
             Delete
           </DropdownMenu.Item>
@@ -354,7 +371,7 @@ export function EntryHeader({editor, editable = true}: EntryHeaderProps) {
                     </DropdownMenu.Trigger>
 
                     <DropdownMenu.Items>
-                      {!isMediaFile && (
+                      {!isMediaFile && !isMediaLibrary && (
                         <DropdownMenu.Item
                           onClick={() => setShowHistory(!showHistory)}
                         >