+++ src/field/link.ts
@@ -1,1 +1,4 @@
-export * from 'alinea/field/link/LinkConstructors'
+export * from 'alinea/field/link/EntryLink'
+export * from 'alinea/field/link/FileLink'
+export * from 'alinea/field/link/ImageLink'
+export * from 'alinea/field/link/Link'